[visionOS] Enable Linear Media Player at build time and runtime https://bugs.webkit.org/show_bug.cgi?id=269767 rdar://123284989
Reviewed by Wenson Hsieh.

Set ENABLE_LINEAR_MEDIA_PLAYER to 1 and enabled the LinearMediaPlayerEnabled runtime setting by
default.

While here, fixed an issue where the wrong mach port was specified in the GPU process seatbelt
profile for using FigVideoTarget rendering and adjusted the imports of WebKitSwift headers to avoid
creating a circular dependency in the Xcode build system.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WTF/wtf/PlatformEnable.h:
* Source/WebKit/Platform/ios/PlaybackSessionInterfaceLMK.mm:
* Source/WebKit/Platform/ios/VideoPresentationInterfaceLMK.mm:
* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.GPU.sb.in:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Source/WebKit/WebKitSwift/LinearMediaKit/WKSLinearMediaPlayer.h:
* Source/WebKit/WebKitSwift/LinearMediaKit/WKSLinearMediaTypes.h:

Canonical link: https://commits.webkit.org/275053@main
