+++ Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.GPU.sb.in
@@ -104,7 +104,7 @@
            (global-name "com.apple.coremedia.capturesource")       ; Also for video capture (<rdar://problem/15794291>).
            (global-name "com.apple.coremedia.cpe.xpc") ; Needed for HDR playback.
 #if ENABLE(LINEAR_MEDIA_PLAYER)
-           (global-name "com.apple.coremedia.videotarget.xpc") ; Needed for FigVideoTarget
+           (global-name "com.apple.coremedia.mediaplaybackd.videotarget.xpc") ; Needed for FigVideoTarget
 #endif
            (global-name "com.apple.coremedia.remotequeue"))
     (allow mach-lookup
